# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LLMLL",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/items/{item_id}/optin": {
      "post": {
        "tags": [
          "queue"
        ],
        "summary": "Optin Item",
        "operationId": "optinItem",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/items/{item_id}/optout": {
      "post": {
        "tags": [
          "queue"
        ],
        "summary": "Optout Item",
        "operationId": "optoutItem",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations/{evaluation_id}/contest": {
      "post": {
        "tags": [
          "contests"
        ],
        "summary": "Contest Evaluation",
        "operationId": "contestEvaluation",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Evaluation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContestResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/attempts/{attempt_id}/contest": {
      "post": {
        "tags": [
          "contests"
        ],
        "summary": "Contest Attempt",
        "description": "Contest an attempt; flashcard attempts from before M4 get their evaluation on demand.",
        "operationId": "contestAttempt",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attempt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContestResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/contests": {
      "get": {
        "tags": [
          "contests"
        ],
        "summary": "List Contests",
        "operationId": "listContests",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestOut"
                  },
                  "title": "Response Listcontests"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/placement/{placement_id}/finish": {
      "post": {
        "tags": [
          "placement"
        ],
        "summary": "Finish Placement",
        "operationId": "finishPlacement",
        "parameters": [
          {
            "name": "placement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Placement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacementFinishOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BandResult": {
        "properties": {
          "n": {
            "type": "integer",
            "title": "N"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "n",
          "score"
        ],
        "title": "BandResult"
      },
...
      "ContestIn": {
        "properties": {
          "item_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Item Ids"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "ContestIn"
      },
      "ContestItemOutcome": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "previous": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "correct",
                  "assisted",
                  "error"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "correct",
              "assisted",
              "error"
            ],
            "title": "Outcome"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "label",
          "previous",
          "outcome"
        ],
        "title": "ContestItemOutcome"
      },
      "ContestOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "evaluation_id": {
            "type": "integer",
            "title": "Evaluation Id"
          },
          "item_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Item Ids"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "resolved"
            ],
            "title": "Status"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "accepted",
                  "rejected",
                  "partial"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "resolver": {
            "type": "string",
            "title": "Resolver"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "replacement_evaluation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replacement Evaluation Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "evaluation_id",
          "item_ids",
          "reason",
          "status",
          "verdict",
          "resolver",
          "rationale",
          "replacement_evaluation_id",
          "created_at",
          "resolved_at"
        ],
        "title": "ContestOut"
      },
      "ContestResultOut": {
        "properties": {
          "contest": {
            "$ref": "#/components/schemas/ContestOut"
          },
          "evaluation_id": {
            "type": "integer",
            "title": "Evaluation Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContestItemOutcome"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "contest",
          "evaluation_id",
          "items"
        ],
        "title": "ContestResultOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OptStatusOut": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "candidate_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Source"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "status",
          "candidate_source"
        ],
        "title": "OptStatusOut"
      },
...
      "PlacementFinishOut": {
        "properties": {
          "placement_id": {
            "type": "string",
            "title": "Placement Id"
          },
          "estimated_level": {
            "type": "string",
            "title": "Estimated Level"
          },
          "previous_level": {
            "type": "string",
            "title": "Previous Level"
          },
          "changed": {
            "type": "boolean",
            "title": "Changed"
          },
          "answered": {
            "type": "integer",
            "title": "Answered"
          },
          "bands": {
            "additionalProperties": {
              "$ref": "#/components/schemas/BandResult"
            },
            "type": "object",
            "title": "Bands"
          }
        },
        "type": "object",
        "required": [
          "placement_id",
          "estimated_level",
          "previous_level",
          "changed",
          "answered",
          "bands"
        ],
        "title": "PlacementFinishOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}